requirements.txt
# Sample requirements for this app
-e git+https://github.com/hedleyroos/django-secretballot.git#egg=django-secretballot
-e git+https://github.com/praekelt/django-likes.git@develop#egg=django-likes
-e git+https://github.com/praekelt/djangorestframework-extras.git@develop#egg=djangorestframework-extras
-e git+https://github.com/praekelt/jmbo.git@feature/django-1.9#egg=jmbo
-e git+https://github.com/praekelt/jmbo-post.git@feature/django-1.9#egg=jmbo-post

Django==1.9.6
django-category==1.9
django-contrib-comments==1.7.1
django-export==1.9
#django-likes==0.2
django-object-tools==1.9
django-pagination-fork==1.0.17
django-photologue==3.5.1
django-preferences==0.0.2
#django-secretballot==0.5.0
django-sites-groups==1.9.2
django-sortedm2m==1.2.2
django-ultracache==1.9.1
ExifRead==2.1.2
#jmbo==2.0.16
Pillow==3.2.0
psycopg2==2.6.1
pytz==2016.4
PyYAML==3.11
requests==2.10.0
djangorestframework==3.3.2
six==1.9.0
